Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-report-to-overrides-report-uri-2.https.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'report-to'.

CONSOLE MESSAGE: Refused to load https://localhost:9443/content-security-policy/support/fail.png because it does not appear in the img-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://localhost:9443/content-security-policy/support/fail.png because it does not appear in the img-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

PASS Test that image does not load
TIMEOUT Event is fired Test timed out
